Add files via upload
Background
Selection of housing is always difficult for someone seeking for a suitable one as it includes various factors and preferences. People prefer to buy a house considering many criteria like- population, quality of life, financial capability, as well as social and natural environments around the housing block.

Dataset
Dataset file name: housing_dataset.csv

Dataset description: Dataset contains total 10 features (columns). It contains the location, housing age, population, number of families in a housing (block), number of rooms, average income of the families in that housing, ocean proximity and other informaiton. Each row indicates a record of a housing block containing the features mentioned earlier.

Features:

latitude (float): Latitude of the location of a housing in conventional geospace
longitude (float): Longitude of the location of a housing in conventional geospace
housing_age (int): Age of the housing in year, the higher number indicates the older housing
total_rooms (int): Total number of rooms in a housing
total_bedrooms (int): Total number of bedrooms in a housing
population (int): Total population of a housing
families (int): Total number of families living in a housing
average_income (float): Average income of the member of a housing in a scale of Tousand Dollar Per Month
ocean_proximity (string): Describing how close the housing is to the ocean
house_value (int): Average individual house price of a housing in Dollers